pyproject.toml
[project]
name = "gridgs-sdk"
version = "0.1.8"
authors = [
  { name="5OKB", email="dev@5okb.com" },
]
description = "Python SDK to simplify integration with GRID services"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "paho-mqtt==2.1.*",
    "python-keycloak==3.*"
]

[project.urls]
"Homepage" = "https://github.com/5OKB/grid-python-sdk"
"Bug Tracker" = "https://github.com/5OKB/grid-python-sdk/issues"
